Added checks if something is already installed

+++ install.sh
@@ -68,13 +68,29 @@ install_all() {
 # Install zsh + oh-my-zsh + powerlevel10k
 install_zsh() {
 	echo "Installing rust..."
-	# install rust
-	curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+  $FILE=/usr/bin/rustup
+  if test -f "$FILE"; then
+    rustup update
+  else 
+    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+  fi
 
 	echo "Installing zsh, oh-my-zsh and powerlevel10k..."
-	pacman -S zsh
-	sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-	git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+  install_packages zsh
+
+  $FILE=/home/$USER/.oh-my-zsh
+  if [ -d "$FILE" ]; then
+    echo "oh-my-zsh already installed..."
+  else
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+  fi
+
+  $FILE=/home/$USER/.oh-my-zsh/custom/themes/powerlevel10k
+  if [ -d "$FILE" ]; then
+    echo "powerlevel10k already installed..."
+  else
+	  git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+  fi
 
 	cp ./.zshrc /home/$USER/
 }
@@ -83,15 +99,16 @@ install_zsh() {
 # Install wezterm
 install_wezterm() {
 	echo "Installing wezterm..."
-	pacman -S wezterm
-	cp ./.wezterm.lua /home/$USER/
+  install_packages wezterm
+
+  cp ./.wezterm.lua /home/$USER/
 }
 
 #=====================================================
 # Install neovim
 install_neovim() {
 	echo "Installing neovim..."
-	pacman -S neovim
+  install_packages neovim
 
 	# required
 	rm /home/$USER/.config/nvim{,.bak}
@@ -102,13 +119,20 @@ install_neovim() {
 	rm /home/$USER/.cache/nvim{,.bak}
 
 	rm -rf /home/$USER/.config/nvim/.git
-	pacman -S lazygit
+  install_packages lazygit
 
 	# for telescope:
-	pacman -S ripgrep fd
+  install_packages ripgrep fd
 
 	echo "Installing lazyvim..."
-	git clone https://github.com/LazyVim/starter /home/$USER/.config/nvim
+  $FILE=/home/$USER/.config/nvim/lua/
+  if [ -d "$FILE" ]; then
+    echo "lazyvim already installed..."
+  else
+    git clone https://github.com/LazyVim/starter /home/$USER/.config/nvim
+  fi
+
+  echo "updating config..."
 	cp -r ./.config/nvim/* /home/$USER/
 }
 
@@ -116,13 +140,23 @@ install_neovim() {
 # doom emacs
 install_dEmacs() {
 	echo "Installing doom emacs"
-	pacman -S emacs
+  install_packages emacs 
 
-	git clone --depth 1 https://github.com/doomemacs/doomemacs /home/$USER/.config/emacs
-	/home/$USER/.config/emacs/bin/doom install
 
-	mkdir /home/$USER/.config/autostart/
-	cp ./emacs.desktop /home/$USER/.config/autostart/.
+  $FILE=/home/$USER/.config/emacs/bin/doom
+  if [ -f "$FILE" ]; then
+    echo "doom emacs already installed..."
+  else
+	  git clone --depth 1 https://github.com/doomemacs/doomemacs /home/$USER/.config/emacs
+    $FILE install
+  fi
+
+  $FILE=/home/$USER/.config/autostart/emacs.desktop
+  if [ -f "$FILE" ]; then
+  else
+	  mkdir /home/$USER/.config/autostart/
+	  cp ./emacs.desktop /home/$USER/.config/autostart/.
+  fi
 
 	doom sync
 }
@@ -131,14 +165,31 @@ install_dEmacs() {
 # hyprland
 install_hyprland() {
 	echo "Installing hyprland..."
-	# yay -S hyprland-git
-	pacman -S sddm dunst pipewire wireplumber xdg-desktop-portal-hyprland polkit-kde-agent qt5-wayland qt6-wayland
+  if pacman -Qs hyprland-git >/dev/null; then
+    echo "hyprland-git is already installed..."
+  else
+    sudo -u $USER yay -S hyprland-git
+  fi
+
+  install_packages sddm dunst pipewire wireplumber xdg-desktop-portal-hyprland polkit-kde-agent qt5-wayland qt6-wayland
 
 	echo "Installing eww..."
 	# eww
 	git clone https://github.com/elkowar/eww
 	cd eww
 	cargo build --release --no-default-features --features=wayland
+}
+
+install_packages() {
+  while [ $# -gt 0 ]; do
+    if pacman -Qs $1 >/dev/null; then
+      echo "$1 is already installed..."
+    else
+      pacman -S $1
+    fi
+    shift
+    ;;
+  done
 }
 
 # Check and install selected components
